promote sector_offset to uint64 in comparison with a uint64 variable
Silences compiler warnings:

dislocker/src/inouts/sectors.c:396:88: warning: comparison of integer expressions of different off_ {long } uint64_ {long unsigned } [-Wsign-compare]
  396 |                 else if(version == V_VISTA && (sector_offset < 16 || sector_offset + 1 == encrypted_volume_total_sectors))
      |                                                                                        ^~
dislocker/src/inouts/sectors.c:401:67: warning: comparison of integer expressions of different off_ {long } uint64_ {long unsigned } [-Wsign-compare]
  401 |                         if(sector_offset < 1 || sector_offset + 1 == encrypted_volume_total_sectors)
      |                                                                   ^~
dislocker/src/inouts/sectors.c: In thread_:
dislocker/src/inouts/sectors.c:480:83: warning: comparison of integer expressions of different off_ {long } uint64_ {long unsigned } [-Wsign-compare]
  480 |                 if(version == V_VISTA && (sector_offset < 16 || sector_offset + 1 == encrypted_volume_total_sectors))
      |                                                                                   ^~
dislocker/src/inouts/sectors.c:485:67: warning: comparison of integer expressions of different off_ {long } uint64_ {long unsigned } [-Wsign-compare]
  485 |                         if(sector_offset < 1 || sector_offset + 1 == encrypted_volume_total_sectors)intaka tand intaka tsignedness: intaka tand intaka tsignedness: encryptfunction intaka tand intaka tsignedness: intaka tand intaka tsignedness:

+++ src/inouts/sectors.c
@@ -393,12 +393,12 @@ static void* thread_decrypt(void* params)
 			);
 			memcpy(loop_output, loop_input, sector_size);
 		}
-		else if(version == V_VISTA && (sector_offset < 16 || sector_offset + 1 == encrypted_volume_total_sectors))
+		else if(version == V_VISTA && (sector_offset < 16 || (uint64_t)sector_offset + 1 == encrypted_volume_total_sectors))
 		{
 			/*
 			 * The firsts sectors are not really encrypted on a Vista volume
 			 */
-			if(sector_offset < 1 || sector_offset + 1 == encrypted_volume_total_sectors)
+                  if(sector_offset < 1 || (uint64_t)sector_offset + 1 == encrypted_volume_total_sectors)
 				fix_read_sector_vista(
 					io_data,
 					loop_input,
@@ -477,12 +477,12 @@ static void* thread_encrypt(void* params)
 		 * NOTE: Seven specificities are dealt with earlier in the process
 		 * see dislocker.c:enlock()
 		 */
-		if(version == V_VISTA && (sector_offset < 16 || sector_offset + 1 == encrypted_volume_total_sectors))
+		if(version == V_VISTA && (sector_offset < 16 || (uint64_t)sector_offset + 1 == encrypted_volume_total_sectors))
 		{
 			/*
 			 * The firsts sectors are not really encrypted on a Vista volume
 			 */
-			if(sector_offset < 1 || sector_offset + 1 == encrypted_volume_total_sectors)
+                  if(sector_offset < 1 || (uint64_t)sector_offset + 1 == encrypted_volume_total_sectors)
 				fix_write_sector_vista(
 					io_data,
 					loop_input,
